feat：使用 gRPC 和 RESTful API 实现搜索服务
- 添加 SearchService 和 SearchQuestions RPC 接口，用于处理问题搜索请求。
- 创建相应的 protobuf 定义，并生成 gRPC 和 HTTP 网关的 Go 代码。
- 为 SearchQuestions 实现 gRPC 服务器处理程序，并与搜索服务逻辑集成。
- 开发了与搜索 API 交互的前端服务，支持身份验证和查询参数。
- 更新数据模型，使用 int64 表示问题相关负载中的 ID，以保持一致性。

+++ cmd/qa-service/internal/service/event_publisher.go
@@ -22,10 +22,10 @@ func (s *qaService) publishQuestionEvent(ctx context.Context, eventType messagin
 			Timestamp: time.Now(),
 		},
 		Payload: messaging.QuestionPayload{
-			ID:         uint64(question.ID),
+			ID:         question.ID,
 			Title:      question.Title,
 			Content:    question.Content,
-			AuthorID:   uint64(question.UserID),
+			AuthorID:   question.UserID,
 			AuthorName: identity.Username,
 			// Tags: question.Tags, // 如果有Tags字段的话
 		},
